pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ollacode"
version = "0.1.0"
description = "경량 CLI 코딩 어시스턴트 — Ollama + Telegram"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
]

[project.scripts]
ollacode = "ollacode.main:main"
